+++ components/members/upload/BasicFormSection.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { Controller, useFormContext } from 'react-hook-form';
 
 import { useGetMemberProfileOfMe } from '@/apiHooks/members';
-import useStringRouterQuery from '@/components/auth/useStringRouterQuery';
+import useStringRouterQuery from '@/hooks/useStringRouterQuery';
 import ImageUploader from '@/components/common/ImageUploader';
 import Input from '@/components/common/Input';
 import FormHeader from '@/components/members/upload/forms/FormHeader';
