+++ pages/auth/callback/facebook/login.tsx
@@ -8,7 +8,7 @@ import { useSetRecoilState } from 'recoil';
 
 import useFacebookAuth from '@/components/auth/identityProvider/useFacebookAuth';
 import { accessTokenAtom } from '@/components/auth/states/accessTokenAtom';
-import useStringRouterQuery from '@/components/auth/useStringRouterQuery';
+import useStringRouterQuery from '@/hooks/useStringRouterQuery';
 import useLastUnauthorized from '@/components/auth/util/useLastUnauthorized';
 import { colors } from '@/styles/colors';
