+++ pages/projects/index.tsx
@@ -2,11 +2,11 @@ import _uniqBy from 'lodash/uniqBy';
 import { FC } from 'react';
 
 import AuthRequired from '@/components/auth/AuthRequired';
-import useStringRouterQuery from '@/components/auth/useStringRouterQuery';
 import Footer from '@/components/common/Footer';
 import Header from '@/components/common/Header';
 import ProjectDetail from '@/components/projects/main/ProjectDetail';
 import ProjectList from '@/components/projects/main/ProjectList';
+import useStringRouterQuery from '@/hooks/useStringRouterQuery';
 import { setLayout } from '@/utils/layout';
 
 const ProjectPage: FC = () => {
